Implemented new features on Deliverable 2: - DataLoader.py: can now ignore duplicates if needed - Split classes and functions that were previously all in TextPreprocessing.py - Added a digit transformer to TextPreprocessor - Added more control to transform method of TextPreprocessor. Now possible to determine if tokenize and if remove stopwords.
